handle when userid is invalid

+++ grails-app/controllers/au/org/ala/alerts/NotificationController.groovy
@@ -115,14 +115,24 @@ class NotificationController {
     @RequireApiKey
     def addMyAlertWS() {
         User user = userService.getUserById(params.userId)
-        notificationService.addAlertForUser(user, Long.valueOf(params.queryId))
-        render([success: true] as JSON)
+        if (user == null) {
+            response.status = 404
+            render ([error : "can't find a user with userId " + params.userId] as JSON)
+        } else {
+            notificationService.addAlertForUser(user, Long.valueOf(params.queryId))
+            render ([success: true] as JSON)
+        }
     }
 
     @RequireApiKey
     def deleteMyAlertWS() {
         User user = userService.getUserById(params.userId)
-        notificationService.deleteAlertForUser(user, Long.valueOf(params.queryId))
-        render([success: true] as JSON)
+        if (user == null) {
+            response.status = 404
+            render ([error : "can't find a user with userId " + params.userId] as JSON)
+        } else {
+            notificationService.deleteAlertForUser(user, Long.valueOf(params.queryId))
+            render ([success: true] as JSON)
+        }
     }
 }
